restapi : ints and shorts overflow
int and short values are overflown before
they even reach the business code, allowing
for wrong values to be set.

When posting to rest api, if the value of an
int is greater than Integer.MAX_VALUE or value
of a short is larger than Short.MAX_VALUE. The
values are overflown when values are unmarshalled
using JAXB.

An excpetion should be raised during JAXB
unmarshalling to indicate that the value is
greater than the MAX_VALUE for the type.

Bug-Url: https://bugzilla.redhat.com/1000796

+++ backend/manager/modules/restapi/types/src/test/java/org/ovirt/engine/api/restapi/types/VmMapperTest.java
@@ -258,7 +258,7 @@ public class VmMapperTest extends
         assertNotNull(cpuTune);
         assertNotNull(cpuTune.getVCpuPin());
         assertEquals(1, cpuTune.getVCpuPin().size());
-        assertEquals(0, cpuTune.getVCpuPin().get(0).getVcpu());
+        assertEquals(0, (int) cpuTune.getVCpuPin().get(0).getVcpu());
         assertEquals("0", cpuTune.getVCpuPin().get(0).getCpuSet());
     }
 
@@ -275,26 +275,26 @@ public class VmMapperTest extends
     @Test()
     public void stringToVCpupinIntervalsList() {
         VCpuPin pin = VmMapper.stringToVCpupin("1#1-4,6");
-        assertEquals(1, pin.getVcpu());
+        assertEquals(1, (int) pin.getVcpu());
         assertEquals("1-4,6", pin.getCpuSet());
     }
 
     @Test()
     public void stringToVCpupinSimple() {
         VCpuPin pin = VmMapper.stringToVCpupin("1#1");
-        assertEquals(1, pin.getVcpu());
+        assertEquals(1, (int) pin.getVcpu());
         assertEquals("1", pin.getCpuSet());
 
         pin = VmMapper.stringToVCpupin("1#10");
-        assertEquals(1, pin.getVcpu());
+        assertEquals(1, (int) pin.getVcpu());
         assertEquals("10", pin.getCpuSet());
 
         pin = VmMapper.stringToVCpupin("1#10,11,12");
-        assertEquals(1, pin.getVcpu());
+        assertEquals(1, (int) pin.getVcpu());
         assertEquals("10,11,12", pin.getCpuSet());
 
         pin = VmMapper.stringToVCpupin("1#10-12,16");
-        assertEquals(1, pin.getVcpu());
+        assertEquals(1, (int) pin.getVcpu());
         assertEquals("10-12,16", pin.getCpuSet());
     }
